% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell profiles function.R
\name{cellProfiles}
\alias{cellProfiles}
\title{Cell Profiles}
\usage{
cellProfiles(data = NULL, position = "center", align = "native",
  reverse = FALSE, contrast = "native", range = c(0.02, 0.98))
}
\arguments{
\item{data}{wide-formated input data consisting of paired columns depicting (1) fluorescent intensity and (2) cell length. Accepts output directly from, eg, ImageJ}

\item{position}{character string (default is "left"), matching either "left" or "center". Determines alignment of stacked profiles}

\item{align}{character string (default is "native"), matching either "native", "orient", or "random", OR to reuse previous alignments, the results of a prior `cellProfiles()` run. Determines orientations of individual profiles.}

\item{reverse}{logical (default is FALSE). Globally reverses alignment of profiles.}

\item{contrast}{character string (default is "native"), matching either "native", "norm", or "max". Controls contrast of individual profiles; "norm" and "max" assist in visualizing profiles with heterogeneous intensity levels}

\item{range}{vector of length 2 (default is c(0.02,0.98) ), defining the lower and upper percentile boundaries of the color scale. Equivalent to adjusting the min/max displayed contrast in ImageJ}
}
\value{
a list of the following values:
\item{interpolated}{processed interpolated data using supplied range values. Columns consist of:
\itemize{x: fluorescence profile x-coordinates}
\itemize{intensity: fluorescence profile value}
\itemize{y: cell number / total cells. Ordered by height for demograph plots}}
\item{native}{as above, but using native resolution}
\item{proportional}{as above, with absolute cell length converted to fraction of cell length}
\item{wide}{wide-formatted table for human eyes; NOT for ggplot2}
\item{fliplist}{a simple vector that records orientations of each cell profile}
}
\description{
\code{cellProfiles} processes fluorescence intensity profiles from, eg, ImageJ, and prepares them for graphing as length-ordered demographs by ggplot
}
\examples{
#calculate cell profiles
data_path <- system.file("extdata", "ftsZ_profiles.csv", package = "cellProfiles")
dtable <- read.table(data_path, header=FALSE, sep=",")
profileResults <- cellProfiles(data=dtable)

#calculate cell profiles, orienting brightest poles on the right & normalizing cell-cell contrast
profileResults <- cellProfiles(data=dtable, align="orient", reverse=TRUE, contrast="norm")

#plot with ggplot
ggplot2::ggplot(data=profileResults$interpolated, mapping=aes(x=x, y=y, fill=intensity, color=intensity) ) + geom_tile() + theme_classic() + scale_y_continuous(expand=c(0.0,0.0), breaks=seq(0,1,0.25), trans="reverse") + scale_fill_gradient(na.value = NA) + scale_color_gradient(na.value=NA) + labs(x="distance from midcell (um)", y="fraction of cell cycle")
}
\seealso{
The cellProfiles vignette for detailed usage instructions:  \code{vignette("cellProfiles")}

\code{\link{cellProfileTruncate}} for trimming input

\code{\link[ggplot2]{ggplot}} for plotting output
}
