% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell profiles function.R
\name{cellProfileTruncate}
\alias{cellProfileTruncate}
\title{Cell Profile Truncate}
\usage{
cellProfileTruncate(data = NULL, adjust = 0)
}
\arguments{
\item{data}{data frame, consisting of paired columns depicting (1) fluorescent intensity and (2) cell length. Accepts output directly from, eg, ImageJ}

\item{adjust}{integer representing the number of data points to trim from each end of the profiles}
}
\value{
a data frame
}
\description{
\code{cellProfileTruncate} takes raw fluorescence intensity profiles from, eg, ImageJ, and truncates the specified number of rows from the beginning and end of each cell
}
\examples{
data_path <- system.file("extdata", "ftsZ_profiles.csv", package = "cellProfiles")
dtable <- read.table(data_path, header=FALSE, sep=",")
dtable_trimmed <- cellProfileTruncate(data=dtable, 1)
}
\seealso{
\code{\link{cellProfiles}}
}
